* add a quantiy field in product to execute the delete low quantiy product
* create a welcome and logout into app.blade.php under layout
* update the ProductController and AdminController for the addition of quantity
* update also the ff. blade template create, edit, user list, edit product create product
* implemet ProductObserver and put a log for create and update and delete

+++ app/Console/Commands/DeleteLowQuantityProducts.php
@@ -26,6 +26,16 @@ class DeleteLowQuantityProducts extends Command
      */
     public function handle()
     {
+
+        $count = Product::where('quantity', '<', 10)->count();
+        if ($count == 0) {
+            $this->info('No low-stock products to delete.');
+            return;
+        }
+
+        // Delete products with quantity less than 10
         Product::where('quantity', '<', 10)->delete();
+
+        $this->info("Deleted {$count} products with low stock.");
     }
 }
